requirements.txt
# UFC Matchmaker — Dependencies

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# UFC API (FritzCapuyan)
ufc_api>=0.0.1

# Data
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Neural Network
torch>=2.1.0

# Database
# (sqlite3 is stdlib)

# CLI / Dashboard
typer>=0.9.0
rich>=13.6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
